(#11156) Fix module path fixtures so rspec works in most rubies
Previously this module was using an experimental pattern using mktmpdir for
creating its module path for rspec-puppet to use, which has since been
deprecated in favour of a static directory. This was largely because ntp was
one of the earlier modules we tried to make work in Jenkins.

This change aligns the module to use the same pattern used in the rspec-
puppet setup stage:

https://github.com/rodjek/rspec-puppet/blob/master/lib/rspec-puppet/setup.rb

This should stop the problems with Ruby 1.8.5 complaining about a lack of
mktmpdir, and also make it more consistent with other modules in the wild.