Added panic parameter to ntp class
This change adds the panic parameter to the ntp class so that it can be
overwritten. The reason for this change is that physical hardware might
also have a large clock skew, not only virtual hardware. In these rare
cases, it would be useful to support enabling: tinker panic 0.

The change to the class still made all the rspec tests pass, so it                                                                                                               should not introduce any regressions. More rspec tests were also added
to exercise the new code paths introduced.