Add new parameters to allow more flexible configuration.
`driftfile` [String]: Allows the setting of the driftfile location.
`keys_enable` [Boolean]: Should the ntp keys option be enabled?
`keys_file` [String]: Where are the keys stored?
`keys_trusted` [Array]: List of keys to trust.
`keys_requestkey` [String]: Request key to use.
`keys_controlkey` [String]: Control key to use.

If keys_enable is set then the directory that keysfile is stored in is
created (recursively).  Currently we don't manage the keys file but that
is an option to be added.