Fix for AIX machines with IPv6 disabled.
This attempts to handle the fact that NTP seems to explode if IPv6
is disabled on AIX, ntp becomes unable to parse the -6 lines that
we rely on in $restrict.  It also apparently predates the KoD filtering
capabilities.  As this stands for 'Kiss of Death' we certainly want to
retain this filtering capability for other distributions and operating
systems.