Enable the ability to set interfaces
Allow the settings of interfaces. If specified, the wildcard
interface will be ignored and the specified interfaces list will be
enabled for listening. This is important on Openstack control nodes
where Neutron can create dozens or even hundreds of interfaces. When new
interfaces are added it causes ntp to rescan. Specificying the interface
list prevents this from occuring.