(MODULES-213) Add support for Solaris
Added Solaris section to params.pp. In Solaris there are two packages
to install to support ntp, SUNWntpr and SUNWntpu. (Root and User
packages.)
Adding two files to the $package_name array caused the install class
to fail so install.pp was updated so that the $package_name was the
resouce title instead of the name attribute.

This change assumes that the packages are already installed. If they
are not then this will fail.