extend Readme for a simple client
these restrictions are needed if you want to install an ntpd that only acts as a client

1) disable global acces
2) allow localhost completly (usefull for just in time config changes via cli)
3) allow minimal access to the source servers

everything for IPv4 and IPv6

settings are based on http://support.ntp.org/bin/view/Support/AccessRestrictions

more and more reflectd NTP attackes are commen on the internet, so it is pretty dangerous to install a simple client without these settings. I guess they could be usefull for other people.

fix typo