Adding ::operatingsystem Amazon
add fedora 21 nodeset

As of this commit, beaker (2.14.1) is currently incompatible with
vagrant/fedora 20+.  See:
https://github.com/dfarrell07/puppet-opendaylight/issues/63#issuecomment-92456469k

(FM-3124) add SLES10 params

(maint) improve workaround for "ssl cert issue"

Solaris 11 doesn't ship the SSL CA root for the forgeapi server
therefore we need to use a different way to deploy the module to
the host. Since the location where puppet looks up the modules
changed between PE 3.x and 2015.2, we need to figure out the correct
location from the host that's being configured.

(maint) improve the ntp/dhcp disintegration to actually do something

If the SUT doesn't have ntp installed at the beginning of the test run
(like ubuntu 12.04) or when the package is removed and re-installed, the
original code would not have deleted the hook script, as it was not
available yet or would have been reinstated. Installing the package
through the class then installed the hook and lead to failures of the
kind this was trying to avoid. using dpkg-divert instead will move the
file away everytime the package is installed by dpkg.

Revert "- add Solaris 12 support"

This reverts commit 158f33fe6098e615ebe91dc562b7c3d61cf06763.

(#2282) Add support for 'disable kernel'

Add support for disabling kernel time discipline in ntp.conf

Update Solaris support for newer Facter

Due to a change in behavior between Facter 1.x and 2.x, the
operatingsystemrelease fact is no longer a desireable determination
of Solaris release.  This work updates the module to use the
kernelrelease fact in place of operatingsystemrelease as it does not
change behavior and is likely a more accurate representation of the
version of Solaris being tested.

Fixed tos_minclock if statement in ntp.conf.etc

Fixed tos lines in ntp_spec.rb