(maint) improve workaround for "ssl cert issue"
Solaris 11 doesn't ship the SSL CA root for the forgeapi server
therefore we need to use a different way to deploy the module to
the host. Since the location where puppet looks up the modules
changed between PE 3.x and 2015.2, we need to figure out the correct
location from the host that's being configured.