(maint) improve the ntp/dhcp disintegration to actually do something
If the SUT doesn't have ntp installed at the beginning of the test run
(like ubuntu 12.04) or when the package is removed and re-installed, the
original code would not have deleted the hook script, as it was not
available yet or would have been reinstated. Installing the package
through the class then installed the hook and lead to failures of the
kind this was trying to avoid. using dpkg-divert instead will move the
file away everytime the package is installed by dpkg.

# Conflicts:
#	spec/spec_helper_acceptance.rb