Update Solaris support for newer Facter
Due to a change in behavior between Facter 1.x and 2.x, the
operatingsystemrelease fact is no longer a desireable determination
of Solaris release.  This work updates the module to use the
kernelrelease fact in place of operatingsystemrelease as it does not
change behavior and is likely a more accurate representation of the
version of Solaris being tested.