Add parameter for interfaces to ignore
The ntp configuration supports some tokens for interfaces that shall be ignored.
Previously the module used the pattern 'wildcard' to indicate that ntp
shouldn't listen to any interface but those specified as listen interfaces.
In some circumstances this is not enough. This commit adds a new parameter
interfaces_ignore that allows to list all interfaces (or interface tokens
like all, ipv6, ipv6 etc.) that should be ignored.