Use config_dir for keys_file
The keys_file defaults to /etc/ntp/keys on many platforms, but the
packages do not create that directory. There is a parameter, config_dir
that can be used to manage this directory, though in our acceptance
tests there was a line that did an mkdir of this directory which was
removed in PR #314 and covered up the fact that the module would not
work by default on many platforms. This should get the tests working
again, and update the debian/ubuntu defaults to their documented
defaults.

Both Debian and Ubuntu actually specify /etc/ntp.keys as the default.
And freebsd, sles, opensuse, and archlinux.

RedHat is the only OS with /etc/ntp/keys as the true default, against
the ntpd standard.