Prepare for puppet4-only release
The following changes will only run (well) on puppet 4.5 or later. Therefore
we can drop the puppet 3 testing on travis, and add more in-depth testing for
puppet 4. Update the version_requirement in metadata.json. We'll also require
the newest version of stdlib for the validate_legacy and deprecation functions.

Since we are dropping compatibilty with puppet 3, users of that version would
experience breakage with the new ntp version, therefore this is a major version
bump.