Adjust for openSUSE Leap and other newer openSUSE releases
Starting with openSUSE 13.2 the ntp service was renamed to "ntpd".

Since openSUSE 13.2 is already out of regular maintenance and older
releases are even more so, simplify the whole logic by only handling
the case for 13.2 and newer releases, which consistently use "ntpd",
and so the case for "ntp" (which was only there for older releases,
but matched for newer releases as well) can be removed alltogether.