(CONT-404) Address set-output deprecation
Prior to this commit, this module contained a set-output command in the
integration_test.yml file. This is a deprecated github command that will
stop working by the end of may.

This commit aims to update the command so that it does not break once
the syntax is fully deprecated.